pyproject.toml
[project]
name = "pwl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi-mail>=1.5.0",
    "fastapi[standard]>=0.116.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "aiosmtpd>=1.4.6",
    "factory-boy>=3.3.3",
    "jsonschema>=4.25.1",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
ignore = ["E731"]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
